requirements.txt
blinker==1.8.2
click==8.1.7
et-xmlfile==1.1.0
Flask==3.0.3
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.24.4
openpyxl==3.1.5
pandas==2.0.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
tzdata==2024.1
Werkzeug==3.0.3
zipp==3.19.2
